+++ test/data/data_preparation.jl
@@ -1,22 +1,11 @@
-## tests.jl
+using DelimitedFiles, Random
+Random.seed!(92906)
 
-#
+data = readdlm("HMP_SRA_gut/HMP_SRA_gut_tiny_meta.tsv", '\t')
 
+three_cat_strcol = vcat("ENV4", rand(["A", "B", "C"], size(data, 1)-1))
+three_cat_intcol = vcat("ENV5", rand([1, 2, 3], size(data, 1)-1))
 
+data_three_cat = hcat(data, three_cat_strcol, three_cat_intcol)
 
-
-
-# conversion to new test results (switch from cauocc to flashweave, but also usable for later updates)
-flashw_test_res(cauocc_test_res) = FlashWeave.Misc.TestResult(cauocc_test_res.stat, cauocc_test_res.pval, cauocc_test_res.df, cauocc_test_res.suff_power)
-
-function convert_exp_dict(exp_dict)
-    new_exp_dict = Dict{String,Any}()
-    for (key, val) in exp_dict
-        if isa(val, Array)
-            new_exp_dict[key] = map(flashw_test_res, val)
-        else
-            new_exp_dict[key] = flashw_test_res(val)
-        end
-    end
-    new_exp_dict
-end
+writedlm("HMP_SRA_gut/HMP_SRA_gut_tiny_meta_oneHotTest.tsv", data_three_cat, '\t')